src/Perun/Offchain/ChannelTxPool: Properly resolve TxInType
The ChainIndex does not return ChainIndexTxs, which contain full TxIn's.
Normally we would assume the ChainIndex to be synced to the latest state
and not withhold any information, but we have to sometimes manually poll
metadata from the ChainIndex itself when fetching the Redeemers which
were used to consume a TxOutRef.